- dialog module fixed - swagger added

+++ dialog-service/src/main/java/ru/skillbox/dialogservice/model/entity/Dialog.java
@@ -17,7 +17,7 @@ public class Dialog {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     @Column(name = "unread_count")
-    private Long unreadCount;
+    private int unreadCount;
     @Column(name = "member1_id")
     private Long member1Id;
     @Column(name = "member2_id")
